requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyMySQL==1.1.2
cryptography==46.0.6
python-dotenv==1.2.2
Werkzeug==3.1.8